==> notebooks/p1_data_preparation.ipynb <==
# Dataset Preparation

## Setup

pwd()

from collections import defaultdict, namedtuple
from os import path, listdir, makedirs
import re

import pandas as pd
import numpy as np

DATA_DIR = path.join(path.pardir, 'data')

PATRIC_AMR = path.join(DATA_DIR, 'PATRIC_genomes_AMR.txt')
GENOME_SUMMARY = path.join(DATA_DIR, 'genome_summary')
GENOME_METADATA = path.join(DATA_DIR, 'genome_metadata')

SAMPLES_DIR = path.join(DATA_DIR, 'samples')

## 1. Samples

### 1.1 Genomes selection

dtypes = {
    'genome_id': str, 
    'genome_name': str,
    # 'taxon_id' 'genome_status'
    'genome_length': np.float64,
    'gc_content': np.float64,
    # 'contig_l50' 'contig_n50' 'chromosomes' 'plasmids' 'contigs'
    # 'patric_cds' 'refseq_cds' 'trna' 'rrnacoarse_consistency'
    'fine_consistency': np.float64,
    'checkm_completeness': np.float64,
    'checkm_contamination': np.float64,
    # 'genome_qualitydate_created' 'date_modified'
}

df = pd.read_csv(
    GENOME_SUMMARY, 
    sep='\t',
    usecols=dtypes.keys(),
    dtype=dtypes
)

def pipe(df, funcs):
    for f in funcs:
        df = f(df)
    return df


def f_mtb(df):
    return df[df.genome_name.str.contains('Mycobacterium tuberculosis')]


def f_gc_content(df):
    # https://bionumbers.hms.harvard.edu/bionumber.aspx?id=101990
    value = 65.6
    lower_bound = value - 0.1
    upper_bound = value + 0.1
    
    return df[(df.gc_content >= lower_bound) & (df.gc_content <= upper_bound)]


def f_genome_length(df):
    # https://bionumbers.hms.harvard.edu/bionumber.aspx?id=101741
    value = 4.41e6
    lower_bound = value - 0.1e6
    upper_bound = value + 0.1e6
    
    return df[(df.genome_length >= lower_bound) & (df.genome_length <= upper_bound)]


def f_qc(df):
    # https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0250092
    # Use more strict QC policy
    contamination_threshold = 10 - 9
    completeness_threshold = 80 + 19
    consistency_threshold = 87 + 11
    
    df = df[df.checkm_contamination <= contamination_threshold]
    df = df[df.checkm_completeness >= completeness_threshold]
    df = df[df.fine_consistency >= consistency_threshold]
    return df


df = pipe(df, [
    f_mtb, 
    f_gc_content, 
    f_genome_length,
    f_qc,
])
df

### 1.2 Genomes metadata enrichment

dtypes = {
    'genome_id': str,
    # 'genome_name' 'organism_name' 
    'taxon_id': str, 
    # 'genome_status'
    'strain': str,
    # 'serovar' 'biovar' 'pathovar' 'mlst' 'other_typing' 'culture_collection' 'type_strain'
    'completion_date': str,
    'publication': str,
    'bioproject_accession': str,
    'biosample_accession': str,
    'assembly_accession': str,
    'genbank_accessions': str,
    'refseq_accessions': str,
    'sequencing_centers': str,
    'sequencing_status': str,
    'sequencing_platform': str,
    'sequencing_depth': str,
    'assembly_method': str,
    'chromosomes': str,
    # 'plasmids'
    'contigs': np.float64,
    'sequences': np.float64,
    # 'genome_length' 'gc_content'
    'patric_cds': np.float64,
    # 'brc1_cds'
    'refseq_cds': np.float64,
    'isolation_site': str,
    'isolation_source': str,
    'isolation_comments': str,
    'collection_date': str,
    # 'isolation_country'
    'geographic_location': str,
    # 'latitude' 'longitude' 'altitude' 'depth' 'other_environmental'
    'host_name': str,
    # 'host_gender' 'host_age' 'host_health' 'body_sample_site' 'body_sample_subsite' 'other_clinical'
    # 'antimicrobial_resistance' 'antimicrobial_resistance_evidence'
    # 'gram_stain' 'cell_shape' 'motility' 'sporulation' 'temperature_range' 'optimal_temperature'
    # 'salinity' 'oxygen_requirement' 'habitat' 'disease'
    'comments': str,
    # 'additional_metadata'
}

df_metadata = pd.read_csv(
    GENOME_METADATA, 
    sep='\t',
    usecols=dtypes.keys(),
    dtype=dtypes
)

df = pd.merge(df, df_metadata, how='inner', on='genome_id')
df

### 1.3 AMR selection and the most common genomes

dtypes = {
    'genome_id': str,
    'antibiotic': str,
    'resistant_phenotype': str,
    'measurement': str,
    'measurement_sign': str,
    'measurement_unit': str,
    'laboratory_typing_method': str,
    'laboratory_typing_method_version': str,
    'laboratory_typing_platform': str,
    'vendor': str,
    'testing_standard': str,
    'testing_standard_year': str,
    'source': str,
}

df_amr = pd.read_csv(
    PATRIC_AMR, 
    sep='\t',
    usecols=dtypes.keys(),
    dtype=dtypes
)

# Selected genomes
df_amr = pd.merge(df[['genome_id', 'genome_name']], df_amr, how='inner', on='genome_id')

# Phenotypes with R/S only
df_amr = df_amr[df_amr.resistant_phenotype.isin(['Resistant', 'Susceptible'])]

# Transform 'laboratory_typing_method_version' in to 'measurement' and 'measurement_unit'
convertions = {}
for t in df_amr[['measurement', 'measurement_unit', 'laboratory_typing_method_version']].itertuples():
    if t.laboratory_typing_method_version is not np.nan:
        m = t.laboratory_typing_method_version.split()
        # assert len(m) > 1, f"Not a measurement: {t}"
        if len(m) > 1:
            convertions[t.Index] = {'measurement': m[0], 'unit': m[1]}
for idx in convertions:
    df_amr.loc[idx, 'measurement'] = convertions[idx]['measurement']
    df_amr.loc[idx, 'measurement_unit'] = convertions[idx]['unit']
    
# Turn 'measurement' column dtype into valid number dtype
extract_number = lambda text: re.findall(r'[-+]?(?:\d*\.*\d)', text)[0]
df_amr['measurement'] = pd.to_numeric(df_amr.measurement.map(extract_number, na_action='ignore'))

# Remove sample without measurement
df_amr = df_amr.dropna(subset=['measurement'])

# Assign NA measurement sign with '='
df_amr['measurement_sign'] = df_amr['measurement_sign'].fillna('=')

# Use only sample in selected genomes
df_amr = pd.merge(df_amr, df['genome_id'], how='inner', on='genome_id')

# Plot the distributions
def phenotype_counts(df_amr):
    dist = defaultdict(int)
    for d in df_amr[['antibiotic', 'resistant_phenotype']].itertuples():
        name = f'{d.antibiotic}:{d.resistant_phenotype.lower()}'
        dist[name] += 1
    
    antibiotics = set([a.split(':')[0] for a in dist.keys()])
    phenotype_counts = [(dist[f'{a}:resistant'], dist[f'{a}:susceptible']) for a in antibiotics]
    return pd.DataFrame(phenotype_counts, index=list(antibiotics), columns=['resistant', 'susceptible'])


phenotype_counts(df_amr)

selected_antibiotics = ['isoniazid', 'rifampin', 'ethambutol']

df_amr = df_amr[df_amr.antibiotic.isin(selected_antibiotics)]

### 1.4 Clinical breakpoints

print(
    df_amr.laboratory_typing_method.unique(),
    df_amr.antibiotic.unique(),
    df_amr.resistant_phenotype.unique(),
    sep='\n\n'
)

# Breakpoints/Critical concentrations tables from CLSI & UCSF
# 
# https://en.wikipedia.org/wiki/Middlebrook_7H9_Broth
# https://en.wikipedia.org/wiki/Middlebrook_7H10_Agar
breakpoints = {
    'ethambutol': {'7H10': 5.0, '7H9': 2.0},
    'isoniazid': {'7H10': 0.2, '7H9': 0.12},
    'rifampin': {'7H10': 1.0,'7H9': 1.0},
}

# Rules:
#   Isolates that cannot grow at ‘critical’ concentrations were then defined as susceptible, 
#   whereas those that can grow were considered resistant
# 
# S = Susceptible
# R = Resistant
# NS = Non-valid-susceptible when the phenotype is S but NOT MIC <= breakpoint
# NR = Non-valid-resistant when the phenotype is R but NOT MIC > breakpoint
def reannotate_phenotype(measurement, antibiotic, phenotype, method):
    def validate_by_beakpoint(antibiotic, media):
        if phenotype == 'Susceptible':
            if not measurement <= breakpoints[antibiotic][media]: 
                return 'NS'
            else:
                return 'S'
        if phenotype == 'Resistant':
            if not measurement > breakpoints[antibiotic][media]:
                return 'NR'
            else:
                return 'R'
    
    if method == 'Agar proportion' or method == 'Agar dilution':
        return validate_by_beakpoint(antibiotic, '7H10')
    if method == 'Broth dilution':
        return validate_by_beakpoint(antibiotic, '7H9')
    
    
assert reannotate_phenotype(2.0, 'ethambutol', 'Resistant', 'Agar proportion') == 'NR' 
assert reannotate_phenotype(1.0, 'rifampin', 'Resistant', 'Agar dilution') == 'NR'
assert reannotate_phenotype(2.0, 'rifampin', 'Resistant', 'Agar dilution') == 'R'
assert reannotate_phenotype(1.0, 'rifampin', 'Susceptible', 'Agar dilution') == 'S'

df_amr['phenotype'] = df_amr.apply(
    lambda r: reannotate_phenotype(
        r['measurement'], 
        r['antibiotic'], 
        r['resistant_phenotype'], 
        r['laboratory_typing_method']
    ),
    axis=1
)

df_amr.value_counts('phenotype')

### 1.6 Only use valid samples (without NR and NS)

df_amr_valid = df_amr[~df_amr.phenotype.isin(['NS', 'NR'])]

print(f'Genomes: {df_amr_valid.genome_id.nunique()}')
print(f'Valid tests: {len(df_amr_valid)}')
phenotype_counts(df_amr_valid)

### 1.7 Remove extreme inconsistence samples

df_amr_valid = df_amr_valid.drop_duplicates(subset=['genome_id', 'antibiotic', 'phenotype'])

print(f'Genomes: {len(df_amr_valid.genome_id.unique())}')
print(f'Valid tests: {len(df_amr_valid)}')
phenotype_counts(df_amr_valid)

def report_inconsistency(df_amr_valid):
    inconsistence_samples = []

    for antibiotic in df_amr_valid.antibiotic.unique():
        checker = defaultdict(list)
    
        for r in df_amr_valid[df_amr_valid.antibiotic == antibiotic].itertuples():
            checker[r.genome_id].append(r.phenotype)
    
        print(f'\nInconsistency in {antibiotic}:')
        for gid in checker.keys():
            if len(checker[gid]) > 1:
                columns = ['genome_id', 'antibiotic', 'phenotype', 'measurement', 'laboratory_typing_method']
                invalid_sample = df_amr_valid.loc[(df_amr.genome_id == gid), columns]
                print(invalid_sample)
                inconsistence_samples.append(gid)
        print('________________________________________________')
            
            
report_inconsistency(df_amr_valid)

bad_samples = ['1324218.3', '1324235.3']

df_amr_valid = df_amr_valid[~df_amr_valid.genome_id.isin(bad_samples)]

report_inconsistency(df_amr_valid)

print(f'Genomes: {len(df_amr_valid.genome_id.unique())}')
print(f'Valid tests: {len(df_amr_valid)}')
phenotype_counts(df_amr_valid)

# Presences of the most common genomes in major antibiotics
# def get_genomes(df_amr_valid, antibiotic):
#     return df_amr_valid[df_amr_valid.antibiotic.str.contains(antibiotic)].genome_id

# g_isoniazid = get_genomes(df_amr_valid, 'isoniazid')
# g_rifampin = get_genomes(df_amr_valid, 'rifampin')
# g_ethambutol = get_genomes(df_amr_valid, 'ethambutol')

# common_genomes = set(g_isoniazid)
# common_genomes.intersection_update(g_rifampin)
# common_genomes.intersection_update(g_ethambutol)

# print(len(common_genomes))

# df_amr_valid = df_amr_valid[df_amr_valid.genome_id.isin(common_genomes)]
# phenotype_counts(df_amr_valid)

df_amr_valid.groupby(['genome_id', 'antibiotic'])['phenotype'].count()

## 2. Results for QC (Mash)

**REMEBER** to add the genome reference!
- Genome ID: **83332.12**
- Genome Name: **Mycobacterium tuberculosis H37Rv**

### 2.1 Export to list and also embed the reference genome

reference_id = '83332.12'

ID_LIST_TXT = path.join(SAMPLES_DIR, '_id_before_mash.txt')
with open(ID_LIST_TXT, 'w') as f:
    gids = [reference_id]
    gids.extend(df_amr_valid.genome_id.unique())
    for gid in gids:
        f.write(f'{gid}\n')

for antibiotic in selected_antibiotics:    
    PANAROO_LIST_TXT = path.join(SAMPLES_DIR, f'_panaroo_{antibiotic}_before_mash.txt')
    gids = [reference_id]
    gids.extend(df_amr_valid[df_amr_valid.antibiotic == antibiotic]['genome_id'].to_list())
    with open(PANAROO_LIST_TXT, 'w') as f:
        for gid in gids:
            f.write(f'{gid}.PATRIC.gff {gid}.fna\n')

### 2.2 Export to Excel

# Export AMR to Excel
wanted_columns = [
    'genome_id',
    'genome_name',
    'antibiotic',
    'phenotype',
    'measurement',
    'measurement_unit',
    'laboratory_typing_method',
]
AMR_XLSX = path.join(DATA_DIR, 'amr_before_mash.xlsx')
df_amr_valid[wanted_columns].to_excel(AMR_XLSX, index=False)

# Export Genomes to Excel
wanted_columns = ['genome_id',
    'genome_name',
    'strain',
    'taxon_id', 
    'genome_length',
    'gc_content',
    'fine_consistency', 
    'checkm_completeness', 
    'checkm_contamination',
    'completion_date',
    'publication',
    'bioproject_accession',
    'biosample_accession',
    'assembly_accession',
    'genbank_accessions',
    'refseq_accessions',
    'sequencing_centers',
    'sequencing_status',
    'sequencing_platform',
    'sequencing_depth',
    'assembly_method',
    'chromosomes',
    'contigs',
    'sequences',
    'patric_cds',
    'refseq_cds',
    'isolation_site',
    'isolation_source',
    'isolation_comments',
    'collection_date',
    'geographic_location',
    'host_name',
    'comments'
]
GENOMES_XLSX = path.join(DATA_DIR, 'genomes_before_mash.xlsx')
df = df[df.genome_id.isin(df_amr_valid.genome_id)]
df[wanted_columns].to_excel(GENOMES_XLSX, index=False)
