==> notebooks/p2_panaroo_preparation.ipynb <==
# Panaroo Preparation

pwd()

from os import path
from collections import defaultdict

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import scipy.stats

DATA_DIR = path.join(path.pardir, 'data')

SAMPLES_DIR = path.join(DATA_DIR, 'samples')
QC_DIR = path.join(DATA_DIR, 'qc')

## 1. Remove Outliers

NGENES_TXT = path.join(QC_DIR, 'ngenes.txt')

genes = pd.read_csv(NGENES_TXT, sep='\t', dtype={'sample': str, 'no_genes': np.float64})
genes.columns = ['genome_id', 'total_genes']
genes

### 1.1 Outliers by Boxplot

upper_outlier = genes[genes['genome_id'] == '1400918.3'].total_genes.to_list()[0]
lower_outlier = genes[genes['genome_id'] == '1773.200'].total_genes.to_list()[0]

genes = genes[(genes.total_genes <= upper_outlier) & (genes.total_genes >= lower_outlier)]
genes

upper: 1400918.3     lower: 1773.200   bad: 1324246.3 1295760.3

## 1.2 Outliers by MDS

mds_outliers = ['1324246.3', '1295760.3']

genes = genes[~genes.genome_id.isin(mds_outliers)]
genes

## 2. Export

### 2.1 Excel stuff: AMR and Genomes

AMR_BEFORE_MASH = path.join(DATA_DIR, 'amr_before_mash.xlsx')
df_amr = pd.read_excel(AMR_BEFORE_MASH, dtype=str)

df_amr = df_amr[df_amr.genome_id.isin(genes.genome_id)]

AMR_READY = path.join(DATA_DIR, 'amr_ready.xlsx')
df_amr.to_excel(AMR_READY, index=False)

df_amr

GENOMES_BEFORE_MASH = path.join(DATA_DIR, 'genomes_before_mash.xlsx')
df_genomes = pd.read_excel(GENOMES_BEFORE_MASH, dtype=str)

df_genomes = df_genomes[df_genomes.genome_id.isin(genes.genome_id)]

GENOMES_READY = path.join(DATA_DIR, 'genomes_ready.xlsx')
df_genomes.to_excel(GENOMES_READY, index=False)

df_genomes

### 2.2 Panaroo stuff

**REMEBER** to add the genome reference again!
- Genome ID: **83332.12**
- Genome Name: **Mycobacterium tuberculosis H37Rv**

reference_id = '83332.12'

ID_LIST_TXT = path.join(SAMPLES_DIR, '_id_ready.txt')
with open(ID_LIST_TXT, 'w') as f:
    gids = [reference_id]
    gids.extend(df_amr.genome_id.unique())
    for gid in gids:
        f.write(f'{gid}\n')

selected_antibiotics = df_amr.antibiotic.unique()
for antibiotic in selected_antibiotics:    
    PANAROO_LIST_TXT = path.join(SAMPLES_DIR, f'_panaroo_{antibiotic}_ready.txt')
    gids = [reference_id]
    gids.extend(df_amr[df_amr.antibiotic == antibiotic]['genome_id'].to_list())
    with open(PANAROO_LIST_TXT, 'w') as f:
        for gid in gids:
            f.write(f'{gid}.PATRIC.gff {gid}.fna\n')

# Plot the distributions
dist = defaultdict(int)
for d in df_amr[['antibiotic', 'phenotype']].itertuples():
    name = f'{d.antibiotic}:{d.phenotype}'
    dist[name] += 1
    
antibiotics = set([a.split(':')[0] for a in dist.keys()])
counts = [(dist[f'{a}:R'], dist[f'{a}:S']) for a in antibiotics]
pd.DataFrame(counts, index=list(antibiotics), columns=['resistant', 'susceptible'])
